host/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
appdirs==1.4.4
asgiref==3.8.1
async-timeout==5.0.1
beautifulsoup4==4.12.3
billiard==4.2.1
build==1.2.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cos-python-sdk-v5==1.9.26
crcmod==1.7
cryptography==44.0.0
Cython==0.29.19
diff-match-patch==20230430
Django==5.0.3
django-cors-headers==4.2.0
django-environ==0.10.0
django-extensions==3.2.3
django-redis==5.0.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
greenlet==3.0.3
idna==3.10
jwcrypto==1.5.6
kombu==5.4.2
lxml==5.3.0
mysqlclient==2.2.7
numpy==1.26.4
openpyxl==3.1.2
pandas==2.1.4
pillow==10.3.0
playwright==1.46.0
prompt_toolkit==3.0.50
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.8.2
pydantic_core==2.20.1
pyee==11.1.0
PyJWT==2.8.0
PyMySQL==1.1.0
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-docx==1.0.1
pytz==2024.2
PyYAML==6.0.1
redis==5.2.0
requests==2.31.0
rest-framework-simplejwt==0.0.2
sh==1.14.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.2
tencentcloud-sdk-python-common==3.0.1056
tencentcloud-sdk-python-tmt==3.0.1056
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.1.0
uWSGI==2.0.22
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
xmltodict==0.14.2
